Transfer Op and Texture

+++ Engine/internal/engine/rendering/vulkan/vulkan_ldevice.h
@@ -50,7 +50,7 @@ public:
 
 	virtual void end_command_buffer_record(GVulkanCommandBuffer* buff) override;
 
-	virtual std::expected<IVulkanBuffer*, VULKAN_BUFFER_CREATION_ERROR> create_buffer(uint64_t size, uint64_t bufferUsageFlag, VmaMemoryUsage memoryUsageFlag ) override;
+	virtual std::expected<IVulkanBuffer*, VULKAN_BUFFER_CREATION_ERROR> create_buffer(uint64_t size, uint32_t bufferUsageFlag, VmaMemoryUsage memoryUsageFlag ) override;
 
 	virtual std::expected< IVulkanImage*, VULKAN_IMAGE_CREATION_ERROR> create_image(const VkImageCreateInfo* imageCreateInfo, VmaMemoryUsage memoryUsageFlag) override;
